requirements.txt
httpx>=0.27.0
PyJWT[crypto]>=2.8.0
